move var manager
