+++ packages/rule-engine/src/manager/variables/db-variable-providers.ts
