+++ packages/rule-engine/src/repositories/index.ts
@@ -1,4 +1,4 @@
 export * from "./db-release-repository.js";
 export * from "./types.js";
-export * from "./variables/types.js";
-export * from "./variables/variables.js";
+export * from "../manager/variables/types.js";
+export * from "../manager/variables/variables.js";
